+++ ck2/cmind/repo.py
@@ -43,7 +43,7 @@ class Repo:
         self.meta = r['meta']
 
         self.path_prefix = self.meta.get('prefix','')
-        if self.path_prefix !='':
+        if self.path_prefix is not None and self.path_prefix.strip() !='':
             self.path_with_prefix = os.path.join(self.path, self.path_prefix)
 
         return {'return':0}